% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazonULC-package.R
\docType{data}
\name{altamira_class}
\alias{altamira_class}
\title{Urban Land Cover for the City of Altamira - Para State, Brazil (year of 2020)}
\format{
A simple features data frame with 228,791 rows and 56 variables:

\describe{
\item{ID}{Unique identifier for each polygon in the data frame.}
\item{ASM_mean}{Average value of the Angular Second Moment (ASM, also called Uniformity). This is a measure of local homogeneity. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{ASM_sum}{Sum value of the Angular Second Moment (ASM, also called Uniformity). This is a measure of local homogeneity. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{B0_median}{Median value of the blue band for all pixels inside the object.}
\item{B0_var}{Variance value of the blue band for all pixels inside the object.}
\item{B1_contr}{Contrast of the green band for all pixels inside the object. It returns a measure of the intensity contrast between a pixel and its southeast neighbor over the object. Contrast is 0 for a constant object.}
\item{B1_dissimi}{Dissimilarity of the green band for all pixels inside the object. It measures how different the elements of the GLCM are from each other and it is high when the local region has a high contrast.}
\item{B1_skewnes}{Skewness value of the green band for all pixels inside the object.}
\item{B1_standar}{Standard Deviation of the green band for all pixels inside the object.}
\item{B2_amplitu}{Amplitude value of the red band for all pixels inside the object. The amplitude means the maximum pixel value minus the minimum pixel value.}
\item{B2_contr}{Contrast of the red band for all pixels inside the object. It returns a measure of the intensity contrast between a pixel and its southeast neighbor over the object. Contrast is 0 for a constant object.}
\item{B2_homogen}{Homogeneity of the red band for all pixels inside the object. It assumes higher values for smaller differences in the GLCM. Also called Inverse Difference Moment. Homogeneity is 1 for a diagonal GLCM.}
\item{B2_max}{Maximum value of the red band for all pixels inside the object.}
\item{B2_min}{Minimum value of the red band for all pixels inside the object.}
\item{B2_skewnes}{Skewness value of the red band for all pixels inside the object.}
\item{B2_standar}{Standard Deviation of the red band for all pixels inside the object.}
\item{B2_var}{Variance value of the red band for all pixels inside the object.}
\item{B3_amplitu}{Amplitude value of the near-infrared band for all pixels inside the object. The amplitude means the maximum pixel value minus the minimum pixel value.}
\item{B3_dissimi}{Dissimilarity of the near-infrared band for all pixels inside the object. It measures how different the elements of the GLCM are from each other and it is high when the local region has a high contrast.}
\item{B3_entropy}{Entropy of the near-infrared band for all pixels inside the object. It measures the disorder in an image. When the image is not uniform, many GLCM elements have small values, resulting in large entropy.}
\item{B3_kurtosi}{Kurtosi of the near-infrared band for all pixels inside the object.}
\item{B3_max}{Maximum value of the near-infrared band for all pixels inside the object.}
\item{B3_min}{Minimun value of the near-infrared band for all pixels inside the object.}
\item{B3_skewnes}{Skewness value of the near-infrared band for all pixels inside the object.}
\item{B3_std}{Standard Deviation of the near-infrared band for all pixels inside the object.}
\item{B3_sum}{Sum value of the near-infrared band for all pixels inside the object.}
\item{B3_var}{Variance value of the near-infrared band for all pixels inside the object.}
\item{BAIN_mean}{Average value of the Bare soil area index (near-infrared version): the normalized division of the blue band by the near-infrared band.}
\item{BAIN_sum}{Sum of the Bare soil area index (near-infrared version): the normalized division of the blue band by the near-infrared band.}
\item{BAI_mean}{Average value of the Bare soil area index (red version): the normalized division of the blue band by the red band.}
\item{BAI_std}{Standard Deviation of the Bare soil area index (red version): the normalized division of the blue band by the red band.}
\item{BAI_sum}{Sum of the Bare soil area index (red version): the normalized division of the blue band by the red band.}
\item{CONTR_min}{Minimum value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 25x25 pixels, isotropically.}
\item{CONTR_std}{Standard Deviation of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 25x25 pixels, isotropically.}
\item{CORR_max}{Maximum value of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{CORR_mean}{Average value of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{CORR_min}{Minimum value of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{CORR_std}{Standard Deviation of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{ENTR_max}{Maximum value of the Entropy. This measure analyses the randomness. It is high when the values of the moving window have similar values. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{ENTR_min}{Minimum value of the Entropy. This measure analyses the randomness. It is high when the values of the moving window have similar values. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{MOC1_max}{Maximum value of the Information Measures of Correlation. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{MOC1_min}{Minimum value of the Information Measures of Correlation. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{MOC1_sum}{Sum value of the Information Measures of Correlation. Calculated over a moving window of 25x25 pixels, isotropically.}
\item{NDVI_mean}{Average value of the Normalized difference vegetation index: the normalized division of the near-infrared band by the red band.}
\item{NDVI_std}{Standard Deviation of the Normalized difference vegetation index: the normalized division of the near-infrared band by the red band.}
\item{NDVI_sum}{Sum value of the Normalized difference vegetation index: the normalized division of the near-infrared band by the red band.}
\item{PBOX_PERIM}{Returns the width of the object's bounding box.}
\item{PCIRCLE}{Relates the areas of the object and the smallest circumscribing circle around the object.}
\item{PDENSITY}{This feature corresponds to the ratio between the polygon area and the polygon radius.	}
\item{PELLIP_FIT}{Finds the minimum circumscribing ellipse to the object and returns the ratio between the object's area and the ellipse area.}
\item{P_FRACDIM}{Returns the fractal dimension of an object.	}
\item{P_PERARAT}{Calculates the ratio between the perimeter and the area of an object.}
\item{P_PERIM}{Returns the perimeter of the object. 	}
\item{TARGET}{Variable related to the sampling process. Null value means that the feature was not selected during sampling.}
\item{CLASS}{The land cover class in which the polygon belongs.}
\item{geometry}{geometry of the simple features.}
}
}
\source{
Vrabel, J.C., Stensaas, G.L., Anderson, C., Christopherson, J., Kim, M., Park, S., and Cantrell, S., 2021, System characterization report on the China-Brazil Earth Resources Satellite-4A (CBERS–4A), chap. J of Ramaseri Chandra, S.N., comp., System characterization of Earth observation sensors: U.S. Geological Survey Open-File Report 2021–1030, 35 p., https://doi.org/10.3133/ofr20211030J.

dos Santos BD, de Pinho CMD, Oliveira GET, Korting TS, Escada MIS, Amaral S. Identifying Precarious Settlements and Urban Fabric Typologies Based on GEOBIA and Data Mining in Brazilian Amazon Cities. Remote Sensing. 2022; 14(3):704. https://doi.org/10.3390/rs14030704

Haralick, R.M., K. Shanmugam, and I. Dinstein (1973). Textural features for image classification. IEEE Transactions on Systems, Man, and Cybernetics, SMC-3(6):610-621.
}
\usage{
data(altamira_class)
}
\description{
A simple features data frame with the features and the land cover classification for Altamira.
In order to obtain this classification, we used the mean-shift algorithm to segment an orthorectified image from the Wide Scan Multispectral and Panchromatic Camera (WPM) of the China-Brazil Earth Resources Satellite (CBERS-4A), dated 07/31/2020 (path: 215, row: 117).
The WMP-CBERS image has 2 m spatial resolution for the panchromatic camera and 8 m for the multispectral camera. After segmentation, we extracted each segment's spectral, Grey-Level Co-occurrence Matrix texture measurements (GLCM), biophysical index, and geometry attributes.
Then, after random and stratified sampling, we classified the segments with the Random Forest machine learning algorithm. The final classification has an accuracy of 88\% to the testing samples (30\% of the sample base).
The classified area overlaps the urban limit of Altamira and is about 152 km². The data is in WGS84 Coordinate Systems, EPSG 4326.
The polygons were classified into:"Shrub Vegetation" (SV), "Herbaceous Vegetation" (HV), "Water" (Wa), "Exposed Ground" (EG), "High Gloss Cover" (HG), "Ceramic Cover" (Ce), "Fiber Cement Cover" (FC), "Asphalt Road" (As), "Terrain Road" (Te), "Cloud" (Cl) and "Shadow" (Sh).
}
\examples{
 data(altamira_class)
 class <- altamira_class$CLASS
}
\keyword{datasets}
