% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazonULC-package.R
\docType{package}
\name{amazonULC-package}
\alias{amazonULC}
\alias{amazonULC-package}
\title{amazonULC: A Data Package with Urban Land Cover Classifications for a Selection of Cities in the Brazilian Amazon}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The amazonULC package aims to provide land cover classification maps of Brazilian Amazonian cities. Land cover maps can contribute to urban planning in Amazonian cities, with applications for monitoring urban sprawl, environmental protection areas, assisting in urban zoning, and identifying areas of high building density, among others. The amazonULC provides a urban land cover classification for the cities of Altamira, Cametá, Marabá, Santarém and part of the Metropolitan Area of Belém, all in the state of Pará, Brazilian Legal Amazon.
}
\references{
\url{https://github.com/dias-bruno/amazonULC}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dias-bruno/amazonULC}
  \item Report bugs at \url{https://github.com/dias-bruno/amazonULC/issues}
}

}
\author{
\strong{Maintainer}: Bruno Dos Santos \email{bruno.santos@inpe.br} (\href{https://orcid.org/0000-0002-6748-2038}{ORCID})

Authors:
\itemize{
  \item Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
}


Bruno Dos Santos, INPE \email{bruno.santos@inpe.br}
}
\keyword{internal}
