% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazonULC-package.R
\docType{data}
\name{cameta_class}
\alias{cameta_class}
\title{Urban Land Cover for the City of Cameta}
\format{
A simple features data frame with 47,834 rows and 28 variables:

\describe{
\item{DN}{Unique identifier for each polygon in the data frame}
\item{BAI..MEAN.}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B3_Mean}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{BAINIR..ME}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{CONTR..MIN}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B2_Standar}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B2_Skewnes}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{ASM..MIN}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B2_Homogen}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{PDENSITY}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B2_Dissim}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B1_Skewnes}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{ENTR..MIN.}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B3_Skewne}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B1_Homogen}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{PCIRCLE}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{VAR..STDDE}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B0_Standar}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B0_Homogen}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{P_FRACDIM}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{B3_Contras}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{CONTR..STD}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{BAI..STDDE}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{P_PERIM}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{BAINIR..ST}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{TARGET}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{CLASSIFICA}{Mean of the DESCRIBE BAI INDEX and UNITS}
\item{geometry}{geometry of the simple features}
}
}
\source{
Honti, Gergely, Tímea Czvetko, and Janos Abonyi. "Data describing the regional Industry 4.0 readiness index." Data in Brief 33 (2020): 106464. \doi{10.1016/j.dib.2020.106464}
}
\usage{
data(cameta_class)
}
\description{
A simple features data frame with . The classification variable was obtained by using a clustering algorithm...
The object is projected using EPSG 32722...UTM Zone 22S...
}
\examples{
 data(cameta_class)
 DN <- cameta_class$DN
}
\keyword{datasets}
