% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazonULC-package.R
\docType{data}
\name{maraba_class}
\alias{maraba_class}
\title{Urban Land Cover for the City of Maraba - Para State, Brazil (year of 2022)}
\format{
A simple features data frame with 232,399 rows and 46 variables:

\describe{
\item{ID}{Unique identifier for each polygon in the data frame.}
\item{ID}{Unique identifier for each polygon in the data frame.}
\item{B0_energy}{Energy of the blue band for all pixels inside the object. Energy returns the squared root of Angular Second Moment, computed by the sum of the squared elements in GLCM. Energy is 1 for a constant image.	}
\item{B0_entropy}{Entropy of the blue band for all pixels inside the object. It measures the disorder in an image. When the image is not uniform, many GLCM elements have small values, resulting in large entropy.}
\item{B2_contr}{Contrast of the red band for all pixels inside the object. It returns a measure of the intensity contrast between a pixel and its southeast neighbor over the object. Contrast is 0 for a constant object.}
\item{B3_contr}{Contrast of the near-infrared band for all pixels inside the object. It returns a measure of the intensity contrast between a pixel and its southeast neighbor over the object. Contrast is 0 for a constant object.}
\item{B3_max}{Maximum value of the near-infrared band for all pixels inside the object.}
\item{B3_min}{Minimun value of the near-infrared band for all pixels inside the object.}
\item{BAI_min}{Minimum value of the Bare soil area index (red version): the normalized division of the blue band by the red band.}
\item{COR_max}{Maximum value of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COR_mean}{Average value of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COR_min}{Minimum value of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COR_range}{Range value of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COT_max}{Maximum value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COT_mean}{Average value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COT_range}{Range value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COT_std}{Standard Deviation value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 3x3 pixels, isotropically.}
\item{DE_range}{Range value of the Difference Entropy. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{DE_std}{Standard Deviation value of the Difference Entropy. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{DV_max}{Maximum value of the Difference Variance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{DV_std}{Standard Deviation value of the Difference Variance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{IDM_max}{Maximum value of the Inverse Difference Moment. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{IDM_mean}{Average value of the Inverse Difference Moment. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{IDM_min}{Minimum value of the Inverse Difference Moment. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{IDM_range}{Range value of the Inverse Difference Moment. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{IFVe_std}{Standard Deviation value of mean value of the vegetation fraction image, derived from a Linear Spectral Mixture Model. }
\item{IFWa_mean}{Average value of mean value of the water fraction image, derived from a Linear Spectral Mixture Model. }
\item{IFWa_sum}{Sum value of mean value of the water fraction image, derived from a Linear Spectral Mixture Model. }
\item{NDRI_std}{Standard Deviation of the Normalized difference roof index: the normalized division of the red band by the blue band.}
\item{NDRI_sum}{Average value of the Normalized difference roof index: the normalized division of the red band by the blue band.}
\item{NDWI_sum}{Sum of the Normalized difference vegetation index: the normalized division of the near-infrared band by the red band.}
\item{PCIRCLE}{Relates the areas of the object and the smallest circumscribing circle around the object.}
\item{PDENSITY}{This feature corresponds to the ratio between the polygon area and the polygon radius.	}
\item{PELLIP_FIT}{Finds the minimum circumscribing ellipse to the object and returns the ratio between the object's area and the ellipse area.}
\item{PGYRATIUS}{This feature is equals the average distance between each vertex of the polygon and it's centroid. The more similar to a circle is the object, the more likely the centroid will be inside it, and therefore this feature will be closer to 0.}
\item{P_FRACDIM}{Returns the fractal dimension of an object.	}
\item{P_PERARAT}{Calculates the ratio between the perimeter and the area of an object.}
\item{P_PERIM}{Returns the perimeter of the object. 	}
\item{Road_perc}{Ratio of the length of roads in each polygon to the area of the polygon. The roads were obtained from open street maps.}
\item{Road_pres}{Presence of roads inside the polygon. The roads were obtained from open street maps.}
\item{SA_range}{Range value of the Sum Avarage. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{SE_std}{Standard Deviation of the Sum Entropy. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{SV_sum}{Sum value of the Sum Variance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{VAR_sum}{Sum value of the Variance. Variance is a measure of gray tone variance within the moving window (second-order moment about the mean). Calculated over a moving window of 3x3 pixels, isotropically.}
\item{TARGET}{Variable related to the sampling process. Null value means that the feature was not selected during sampling.}
\item{CLASS}{The land cover class in which the polygon belongs.}
\item{geometry}{geometry of the simple features.}
}
}
\source{
Vrabel, J.C., Stensaas, G.L., Anderson, C., Christopherson, J., Kim, M., Park, S., and Cantrell, S., 2021, System characterization report on the China-Brazil Earth Resources Satellite-4A (CBERS–4A), chap. J of Ramaseri Chandra, S.N., comp., System characterization of Earth observation sensors: U.S. Geological Survey Open-File Report 2021–1030, 35 p., https://doi.org/10.3133/ofr20211030J.

dos Santos BD, de Pinho CMD, Oliveira GET, Korting TS, Escada MIS, Amaral S. Identifying Precarious Settlements and Urban Fabric Typologies Based on GEOBIA and Data Mining in Brazilian Amazon Cities. Remote Sensing. 2022; 14(3):704. https://doi.org/10.3390/rs14030704

Haralick, R.M., K. Shanmugam, and I. Dinstein (1973). Textural features for image classification. IEEE Transactions on Systems, Man, and Cybernetics, SMC-3(6):610-621.
}
\usage{
data(maraba_class)
}
\description{
A simple features data frame with the features and the land cover classification for Marabá.
In order to obtain this classification, we used the mean-shift algorithm to segment an orthorectified image from the Wide Scan Multispectral and Panchromatic Camera (WPM) of the China-Brazil Earth Resources Satellite (CBERS-4A), dated 09/03/2022 (path: 211, row: 120).
The WMP-CBERS image has 2 m spatial resolution for the panchromatic camera and 8 m for the multispectral camera. After segmentation, we extracted each segment's spectral, Grey-Level Co-occurrence Matrix texture measurements (GLCM), biophysical index, and geometry attributes.
Then, after random and stratified sampling, we classified the segments with the Random Forest machine learning algorithm. The final classification has an accuracy of 93\% to the testing samples (30\% of the sample base).
The classified area overlaps the urban limit of Maraba and is about 425 km². The data is in WGS84 Coordinate Systems, EPSG 4326.
The polygons were classified into:"Shrub Vegetation" (SV), "Herbaceous Vegetation" (HV), "Water" (Wa), "Exposed Ground" (EG), "High Gloss Cover" (HG), "Ceramic Cover" (Ce), "Fiber Cement Cover" (FC), "Asphalt Road" (As), "Terrain Road" (Te), "Cloud" (Cl) and "Shadow" (Sh).
}
\examples{
 data(maraba_class)
 class <- maraba_class$CLASS
}
\keyword{datasets}
