% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazonULC-package.R
\docType{data}
\name{santarem_class}
\alias{santarem_class}
\title{Urban Land Cover for the City of Santarem - Para State, Brazil (year of 2020)}
\format{
A simple features data frame with 65,822 rows and 33 variables:

\describe{
\item{ID}{Unique identifier for each polygon in the data frame.}
\item{ASM_mean}{Average value of the Angular Second Moment (ASM, also called Uniformity). This is a measure of local homogeneity. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{B0_amplitu}{Amplitude value of the blue band for all pixels inside the object. The amplitude means the maximum pixel value minus the minimum pixel value.}
\item{B0_var}{Variance value of the blue band for all pixels inside the object.}
\item{B3_mean}{Average value of the near-infrared band for all pixels inside the object.}
\item{B3_min}{Minimun value of the near-infrared band for all pixels inside the object.}
\item{BAI_var}{Variance of the Bare soil area index (red version): the normalized division of the blue band by the red band.}
\item{COR_sum}{Sum value of the Correlation. This measure analyses the linear dependency of grey levels of neighboring pixels. Typically high, when the scale of local texture is larger than the distance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COT_max}{Maximum value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COT_mean}{Average value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 3x3 pixels, isotropically.}
\item{COT_var}{Variance value of the Constrast. This measure analyses the image contrast (locally gray-level variations) as the linear dependency of grey levels of neighboring pixels (similarity). Calculated over a moving window of 3x3 pixels, isotropically.}
\item{DE_mean}{Average value of the Difference Entropy. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{IDM_mean}{Average value of the Inverse Difference Moment. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{IDM_min}{Minimum value of the Inverse Difference Moment. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{IFAG_mean}{Average value of mean value of the water fraction image, derived from a Linear Spectral Mixture Model. }
\item{IFAG_var}{Variance value of mean value of the water fraction image, derived from a Linear Spectral Mixture Model. }
\item{IFSE_mean}{Average value of mean value of the exposed soil fraction image, derived from a Linear Spectral Mixture Model. }
\item{IFSE_min}{Minimim value of mean value of the exposed soil fraction image, derived from a Linear Spectral Mixture Model. }
\item{IFVE_mean}{Average value of mean value of the vegetation fraction image, derived from a Linear Spectral Mixture Model. }
\item{MC1_mean}{Average value of the Information Measures of Correlation. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{MC1_min}{Minimum value of the Information Measures of Correlation. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{NDRI_sum}{Average value of the Normalized difference roof index: the normalized division of the red band by the blue band.}
\item{NDRI_var}{Variance value of the Normalized difference roof index: the normalized division of the red band by the blue band.}
\item{NDVI_var}{Variance value of the Normalized difference vegetation index: the normalized division of the near-infrared band by the red band.}
\item{Road_perc}{Ratio of the length of roads in each polygon to the area of the polygon. The roads were obtained from open street maps.}
\item{Road_pres}{Presence of roads inside the polygon. The roads were obtained from open street maps.}
\item{SA_min}{Minimum value of the Sum Avarage. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{SA_var}{Variance value of the Sum Avarage. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{SE_mean}{Average value of the Sum Entropy. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{SV_mean}{Average value of the Sum Variance. Calculated over a moving window of 3x3 pixels, isotropically.}
\item{TARGET}{Variable related to the sampling process. Null value means that the feature was not selected during sampling.}
\item{CLASS}{The land cover class in which the polygon belongs.}
\item{geometry}{geometry of the simple features.}
}
}
\source{
Vrabel, J.C., Stensaas, G.L., Anderson, C., Christopherson, J., Kim, M., Park, S., and Cantrell, S., 2021, System characterization report on the China-Brazil Earth Resources Satellite-4A (CBERS–4A), chap. J of Ramaseri Chandra, S.N., comp., System characterization of Earth observation sensors: U.S. Geological Survey Open-File Report 2021–1030, 35 p., https://doi.org/10.3133/ofr20211030J.

dos Santos BD, de Pinho CMD, Oliveira GET, Korting TS, Escada MIS, Amaral S. Identifying Precarious Settlements and Urban Fabric Typologies Based on GEOBIA and Data Mining in Brazilian Amazon Cities. Remote Sensing. 2022; 14(3):704. https://doi.org/10.3390/rs14030704

Haralick, R.M., K. Shanmugam, and I. Dinstein (1973). Textural features for image classification. IEEE Transactions on Systems, Man, and Cybernetics, SMC-3(6):610-621.
}
\usage{
data(santarem_class)
}
\description{
A simple features data frame with the features and the land cover classification for Santarem
In order to obtain this classification, we used the mean-shift algorithm to segment an orthorectified image from the Wide Scan Multispectral and Panchromatic Camera (WPM) of the China-Brazil Earth Resources Satellite (CBERS-4A), dated 11/09/2020 (path: 219, row: 116).
The WMP-CBERS image has 2 m spatial resolution for the panchromatic camera and 8 m for the multispectral camera. After segmentation, we extracted each segment's spectral, Grey-Level Co-occurrence Matrix texture measurements (GLCM), biophysical index, and geometry attributes.
Then, after random and stratified sampling, we classified the segments with the Random Forest machine learning algorithm. The final classification has an accuracy of 98\% to the testing samples (30\% of the sample base).
The classified area overlaps the urban limit of Santarem and is about 143 km². The data is in WGS84 Coordinate Systems, EPSG 4326.
The polygons were classified into:"Shrub Vegetation" (SV), "Herbaceous Vegetation" (HV), "Water" (Wa), "Exposed Ground" (EG), "High Gloss Cover" (HG), "Ceramic Cover" (Ce), "Fiber Cement Cover" (FC), "Asphalt Road" (As), "Terrain Road" (Te), "Cloud" (Cl) and "Shadow" (Sh).
}
\examples{
 data(santarem_class)
 class <- santarem_class$CLASS
}
\keyword{datasets}
