% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazonULC-package.R
\docType{package}
\name{amazonULC-package}
\alias{amazonULC}
\alias{amazonULC-package}
\title{amazonULC: A Data Package with Urban Land Cover Classifications for a Selection of Cities in the Brazilian Amazon}
\description{
What the package does (one paragraph).
}
\references{
\url{https://github.com/dias-bruno/amazonULC}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dias-bruno/amazonULC}
  \item \url{https://dias-bruno.github.io/amazonULC/}
  \item Report bugs at \url{https://github.com/dias-bruno/amazonULC/issues}
}

}
\author{
\strong{Maintainer}: Bruno Dos Santos \email{bruno.santos@inpe.br} (\href{https://orcid.org/0000-0002-6748-2038}{ORCID})

Authors:
\itemize{
  \item Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
}


Bruno Dos Santos, INPE \email{bruno.santos@inpe.br}

Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
\keyword{internal}
